greeting-fastmcp/pyproject.toml
[project]
name = "greeting-fastmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.13.0.2",
    "python-dotenv>=1.0.0"
]
